move core services modules from scripts into cgse_core
