+++ libs/cgse-core/src/cgse_core/_start.py
