+++ libs/cgse-core/src/cgse_core/_status.py
