+++ libs/cgse-core/src/cgse_core/_stop.py
